refactor(web): one slug-set store instead of three
Catalogue #38. savedJobs, viewedJobs and dismissedJobs already shared the
load-once/reset scaffolding through UserResource<T>. What stayed duplicated was
the payload half — a SvelteSet in $state, has, mark, sometimes unmark, and the
two hooks that swap the set on load and drop it on reset. Thirty lines, three
times, differing by one method call.

SlugSet owns that; a subclass supplies load() and nothing else, so each store is
five lines. unmark goes on the base rather than into the two subclasses that had
it: viewedJobs simply had no caller for it, and an absent caller is not a
different rule.

No unit test, and the reason is the repo's own: SlugSet is all runes, and
paginated.svelte.test.ts already records that "$state fields need a Svelte
runtime this test env doesn't provide". The collapse was verified by reading the
three bodies side by side first, and pnpm run check stays at its baseline.

Checked while here: catalogue #36 ("the single source of display labels is
forked") is already closed by S1 — one titleCase, one categoryLabel, one
RELOCATION_LABELS, imported by both consumers, with a test. The catalogue's
numbering and the shortlist's ids do not line up, so overlap has to be checked in
the code rather than inferred.
