+++ web/src/lib/savedJobs.svelte.ts
@@ -12,42 +12,12 @@
 // and the set stays empty for signed-out users. A failed load leaves the set empty —
 // nothing shows filled, the correct degraded state.
 
-import { SvelteSet } from 'svelte/reactivity';
 import { api } from '$lib/api';
-import { UserResource } from '$lib/userResource.svelte';
+import { SlugSet } from '$lib/userResource.svelte';
 
-class SavedJobs extends UserResource<string[]> {
-  // SvelteSet (not a plain Set): a plain Set in $state is not deeply reactive, so
-  // an in-place `.add`/`.delete` would not re-run readers. SvelteSet makes both the
-  // mutation and the load reassignment trigger dependent $derived/$effect (e.g.
-  // JobRow's `saved`).
-  #slugs = $state(new SvelteSet<string>());
-
-  has(slug: string): boolean {
-    return this.#slugs.has(slug);
-  }
-
-  /** Mark a slug saved locally (e.g. right after a successful save), so its card's
-   *  bookmark fills immediately without re-fetching the whole set. */
-  mark(slug: string) {
-    this.#slugs.add(slug);
-  }
-
-  /** Clear a slug's saved mark locally (e.g. right after a successful unsave). */
-  unmark(slug: string) {
-    this.#slugs.delete(slug);
-  }
-
+class SavedJobs extends SlugSet {
   protected load(): Promise<string[]> {
     return api.listSavedSlugs();
-  }
-
-  protected apply(slugs: string[]) {
-    this.#slugs = new SvelteSet(slugs);
-  }
-
-  protected clearState() {
-    this.#slugs = new SvelteSet();
   }
 }
